finally done with player movement; decided to go with hardcode because trigonometry would create some edge cases that would end up in some ifs too

+++ proj/src/models/Entity.h
@@ -24,7 +24,7 @@ typedef struct {
     Position* position;
     HurtBox* hurtBox;
     double velocity;        // pixels/s
-    uint16_t direction;     // [0,360[, 0 being right
+    int direction;     // [0,360[, 0 being right
     EntityType type;  
     void* typeInfo;         // entity type specific info
 } Entity;
